# caliborn.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "caliborn",
    "description": "API server for lumiRadio",
    "contact": {
      "name": "cozygalvinism"
    },
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/admin/users/{id}": {
      "get": {
        "tags": [
          "crate::routes::admin::users"
        ],
        "operationId": "get_user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "user_jwt": []
          },
          {
            "user_api_key": []
          }
        ]
      },
      "patch": {
        "tags": [
          "crate::routes::admin::users"
        ],
        "operationId": "patch_user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "user_jwt": []
          },
          {
            "user_api_key": []
          }
        ]
      }
    },
...
    "/songs/current": {
      "get": {
        "tags": [
          "crate::routes::songs"
        ],
        "operationId": "get_currently_playing",
        "responses": {
          "200": {
            "description": "The song currently on air",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SongDto"
                }
              }
            }
          },
          "404": {
            "description": "Nothing has been played yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "An internal server error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/songs/favourite": {
      "post": {
        "tags": [
          "crate::routes::songs"
        ],
        "operationId": "mark_song_as_favourite",
        "parameters": [
          {
            "name": "song_id",
            "in": "query",
            "description": "File hash of the song to favourite",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Song added to the caller's favourites"
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Caller lacks the `use_bot` permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No song matches the given file hash",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "An internal server error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "user_jwt": []
          },
          {
            "user_api_key": []
          }
        ]
      },
      "delete": {
        "tags": [
          "crate::routes::songs"
        ],
        "operationId": "unmark_song_as_favourite",
        "parameters": [
          {
            "name": "song_id",
            "in": "query",
            "description": "File hash of the song to un-favourite",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Song removed from the caller's favourites"
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Caller lacks the `use_bot` permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No song matches the given file hash",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "An internal server error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "user_jwt": []
          },
          {
            "user_api_key": []
          }
        ]
      }
    },
    "/songs/favourite/current": {
      "post": {
        "tags": [
          "crate::routes::songs"
        ],
        "operationId": "mark_currently_playing_song_as_favourite",
        "responses": {
          "200": {
            "description": "The song currently on air was added to the caller's favourites"
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Caller lacks the `use_bot` permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Nothing has been played yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "An internal server error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "user_jwt": []
          },
          {
            "user_api_key": []
          }
        ]
      }
    },
    "/songs/favourites": {
      "get": {
        "tags": [
          "crate::routes::songs"
        ],
        "operationId": "search_favourite_songs",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter[artist]",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter[album]",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter[title]",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results restricted to the caller's favourites",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_SongDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Caller lacks the `use_bot` permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "An internal server error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "user_jwt": []
          },
          {
            "user_api_key": []
          }
        ]
      }
    },
    "/songs/history": {
      "get": {
        "tags": [
          "crate::routes::songs"
        ],
        "operationId": "get_song_history",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Recently played songs, newest first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_SongDto"
                }
              }
            }
          },
          "500": {
            "description": "An internal server error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/songs/queue": {
      "get": {
        "tags": [
          "crate::routes::songs"
        ],
        "operationId": "get_request_queue",
        "responses": {
          "200": {
            "description": "Pending song requests, in play order",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SongListDto"
                }
              }
            }
          },
          "500": {
            "description": "An internal server error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/songs/request": {
      "post": {
        "tags": [
          "crate::routes::songs"
        ],
        "operationId": "request_song",
        "parameters": [
          {
            "name": "file_hash",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Song queued; includes the user and song cooldowns now in effect",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SongWithCooldownInfo"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Caller lacks the `use_bot` permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No song matches the given file hash",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "User or song is still on cooldown",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "An internal server error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "user_jwt": []
          },
          {
            "user_api_key": []
          }
        ]
      }
    },
    "/songs/search": {
      "get": {
        "tags": [
          "crate::routes::songs"
        ],
        "operationId": "search_song",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter[artist]",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter[album]",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter[title]",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Full-text search results over the song library",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_SongDto"
                }
              }
            }
          },
          "500": {
            "description": "An internal server error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdminUserDto": {
        "type": "object",
        "required": [
          "id",
          "boonbucks",
          "watched_time",
          "migrated",
          "role",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "boonbucks": {
            "type": "integer",
            "format": "int32"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "last_message_sent": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "migrated": {
            "type": "boolean"
          },
          "role": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "username": {
            "type": [
              "string",
              "null"
            ]
          },
          "watched_time": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "An error response containing details about the problem",
        "required": [
          "message",
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "A machine-readable error code identifying the error type.\n\nCommon values include \"Bad Request\", \"Unauthorized\", \"Internal Server Error\",\nand \"Not Found\".",
            "examples": [
              "Bad Request",
              "Unauthorized",
              "Internal Server Error"
            ]
          },
          "message": {
            "type": "string",
            "description": "A human-readable message describing the error.",
            "examples": [
              "Invalid authorization code"
            ]
          }
        },
        "examples": [
          {
            "error": "Unauthorized",
            "message": "Invalid authorization code"
          }
        ]
      },
...
      "Page_SongDto": {
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "total_pages"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id",
                "title",
                "artist",
                "album",
                "duration",
                "bitrate"
              ],
              "properties": {
                "album": {
                  "type": "string"
                },
                "artist": {
                  "type": "string"
                },
                "bitrate": {
                  "type": "integer",
                  "format": "int32"
                },
                "duration": {
                  "type": "number",
                  "format": "double"
                },
                "id": {
                  "type": "string"
                },
                "title": {
                  "type": "string"
                }
              }
            }
          },
          "page": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "page_size": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "total_pages": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "SongDto": {
        "type": "object",
        "required": [
          "id",
          "title",
          "artist",
          "album",
          "duration",
          "bitrate"
        ],
        "properties": {
          "album": {
            "type": "string"
          },
          "artist": {
            "type": "string"
          },
          "bitrate": {
            "type": "integer",
            "format": "int32"
          },
          "duration": {
            "type": "number",
            "format": "double"
          },
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "SongListDto": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/SongDto"
        }
      },
      "SongWithCooldownInfo": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SongDto"
          },
          {
            "$ref": "#/components/schemas/CooldownInfo"
          }
        ]
      },
...
      "UpdateUserRequest": {
        "type": "object",
        "properties": {
          "boonbucks": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "migrated": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "username": {
            "type": [
              "string",
              "null"
            ]
          },
          "watched_time": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          }
        }
      },
...
    },
    "securitySchemes": {
      "user_api_key": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key"
      },
      "user_jwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}